minecraft/service/environment.go: Moved Environment from dial.go
